+++ OPHD/UI/FactoryListBox.cpp
@@ -137,7 +137,7 @@ void FactoryListBox::currentSelection(Factory* f)
 		FactoryListBoxItem* item = static_cast<FactoryListBoxItem*>(mItems[i]);
 		if (item->factory == f)
 		{
-			setSelection(static_cast<int>(i));
+			setSelection(i);
 			return;
 		}
 	}
@@ -168,8 +168,8 @@ void FactoryListBox::update()
 		drawItem(renderer, *static_cast<FactoryListBoxItem*>(mItems[i]),
 			positionX(),
 			positionY() + (static_cast<int>(i) * LIST_ITEM_HEIGHT),
-			item_width(),
-			draw_offset(),
+			static_cast<int>(item_width()),
+			static_cast<int>(draw_offset()),
 			i == currentSelection());
 	}
